Don't enable never_type on docs.rs
Now that the impl for Result<T, !> subsumes the impl for
Result<T, Infallible> in nightly Rust, the documentation for stable
compiler users would become misleading.
With Result::into_ok soon to be merged, this crate's function shifts
to providing support for pre-1.41 Rust library API, and the
documentation should be geared towards that.